requirements/polaris_gc_requirements.txt
globus-compute-endpoint==2.2.4
numpy==1.24.4
matplotlib==3.7.2
hyperspy==1.7.5
ultralytics==8.0.20
imageio-ffmpeg==0.4.8